requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0
djangorestframework==3.9.0
python-decouple==3.1
pytz==2018.5
static3==0.7.0
gunicorn
psycopg2